solutions/string_permutations.py
from itertools import permutations

def find_permutations(s):
    """
    Generates all permutations of the characters in the input string.

    Parameters:
        s (str): The input string.

    Returns:
        list: A list containing all unique permutations of the string.

    Raises:
        TypeError: If the input is not a string.

    Example:
        >>> find_permutations("abc")
        ['abc', 'acb', 'bac', 'bca', 'cab', 'cba']
    """
    if not isinstance(s, str):
        raise TypeError("Input must be a string")
    if s == "":
        return [""]
    return sorted(set("".join(p) for p in permutations(s)))
